Clean up the source of MechanicsValues.
